Consolidates kernelvethpair and kerneltapv2 under a single kernel.
Turned out this was the fastest cleanest simplest way to transition
kerneltapv2 to using sendfd/recvfd
