+++ pkg/networkservice/mechanisms/kernel/common.go
@@ -14,21 +14,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kernelvethpair
+package kernel
 
 import (
 	"github.com/networkservicemesh/api/pkg/api/networkservice/mechanisms/kernel"
 	"go.ligato.io/vpp-agent/v3/proto/ligato/configurator"
+	"go.ligato.io/vpp-agent/v3/proto/ligato/linux"
 	linuxinterfaces "go.ligato.io/vpp-agent/v3/proto/ligato/linux/interfaces"
 	linuxnamespace "go.ligato.io/vpp-agent/v3/proto/ligato/linux/namespace"
 	vppinterfaces "go.ligato.io/vpp-agent/v3/proto/ligato/vpp/interfaces"
 )
 
 const (
-	fileScheme = "file"
+	// MECHANISM string
+	MECHANISM = kernel.MECHANISM
+
+	fileScheme   = "file"
+	vnetFilename = "/dev/vhost-net"
 )
 
-func appendInterfaceConfig(conf *configurator.Config, name, ifaceName, netnsFilename string) {
+func vethPairTemplate(conf *configurator.Config, name, ifaceName, netnsFilename string) {
 	conf.GetLinuxConfig().Interfaces = append(conf.GetLinuxConfig().Interfaces,
 		&linuxinterfaces.Interface{
 			Name:       name + "-veth",
@@ -67,6 +72,40 @@ func appendInterfaceConfig(conf *configurator.Config, name, ifaceName, netnsFile
 		Link: &vppinterfaces.Interface_Afpacket{
 			Afpacket: &vppinterfaces.AfpacketLink{
 				HostIfName: linuxIfaceName(name),
+			},
+		},
+	})
+}
+
+func tapV2Template(conf *configurator.Config, name, ifaceName, netnsFilename string) {
+	// We append an Interfaces.  Interfaces creates the vpp side of an interface.
+	//   In this case, a Tapv2 interface that has one side in vpp, and the other
+	//   as a Linux kernel interface
+	conf.GetVppConfig().Interfaces = append(conf.GetVppConfig().GetInterfaces(), &vppinterfaces.Interface{
+		Name:    name,
+		Type:    vppinterfaces.Interface_TAP,
+		Enabled: true,
+		Link: &vppinterfaces.Interface_Tap{
+			Tap: &vppinterfaces.TapLink{
+				Version: 2,
+			},
+		},
+	})
+	// We apply configuration to LinuxInterfaces
+	// Important details:
+	//    - LinuxInterfaces.HostIfName - must be no longer than 15 chars (linux limitation)
+	conf.GetLinuxConfig().Interfaces = append(conf.GetLinuxConfig().Interfaces, &linux.Interface{
+		Name:       name,
+		Type:       linuxinterfaces.Interface_TAP_TO_VPP,
+		Enabled:    true,
+		HostIfName: linuxIfaceName(ifaceName),
+		Namespace: &linuxnamespace.NetNamespace{
+			Type:      linuxnamespace.NetNamespace_FD,
+			Reference: netnsFilename,
+		},
+		Link: &linuxinterfaces.Interface_Tap{
+			Tap: &linuxinterfaces.TapLink{
+				VppTapIfName: name,
 			},
 		},
 	})
