+++ pkg/networkservice/mechanisms/kernel/common_test.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kernelvethpair_test
+package kernel_test
 
 import (
 	"fmt"
